% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_management.R
\name{assign_final}
\alias{assign_final}
\title{Assign Final Value}
\usage{
assign_final(name, value, env = parent.frame(), verbose = TRUE)
}
\arguments{
\item{name}{Variable name to assign value to.}

\item{value}{Value assigned.}

\item{env}{Environment in which to conduct the assignment. Defaults to the calling environment.}

\item{verbose}{Default TRUE. Whether to print warning message if a parameter has already been assigned.}
}
\value{
No value is returned.
}
\description{
Helper function. Assigns an immutable value to a variable name within a specified
environment. This function is meant to be used in the context of setting processing
batch parameters.
}
