% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_management.R
\name{checkArgsAgainstBatchParams}
\alias{checkArgsAgainstBatchParams}
\title{Check Function Arguments against Batch-Specific Parameters}
\usage{
checkArgsAgainstBatchParams(override = FALSE, ...)
}
\value{

}
\description{
If a processing batch is currently set, inserting this function into another function from
neonMicrobe will check the parent function's arguments against the parameters associated
with the current processing batch. If override=TRUE, the current batch's parameters would
take precedence. For example, rather than writing to the standard outputs directory, the
parent function would write to the current batch's outputs directory. Rather than using
quality filtering parameters defined on-the-fly, the parent function would use the quality
filtering parameters associated with the current batch. Override does not apply to
parameters that are inconsequential to the resulting outputs, like VERBOSE and MULTITHREAD.
}
