% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapseNoMismatch2}
\alias{collapseNoMismatch2}
\title{Combine together sequences that are identical up to shifts and/or length (with progress bar)}
\usage{
collapseNoMismatch2(
  seqtab,
  minOverlap = 20,
  orderBy = "abundance",
  identicalOnly = FALSE,
  vec = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{seqtab}{(Required). A sample by sequence matrix, the return of \code{\link[dada2]{makeSequenceTable}}.}

\item{minOverlap}{(Optional). numeric(1). Default 20. The minimum amount of overlap between sequences required to collapse them together.}

\item{orderBy}{(Optional). character(1). Default "abundance". Specifies how the sequences (columns) of the returned table should be ordered (decreasing). Valid values: "abundance", "nsamples", NULL.}

\item{identicalOnly}{(Optional). logical(1). Default FALSE. If TRUE, only identical sequences (i.e. duplicates) are collapsed together.}

\item{vec}{(Optional). logical(1). Default TRUE. Use the vectorized aligner. Should be turned off if sequences exceed 2kb in length.}

\item{verbose}{(Optional). logical(1). Default FALSE. If TRUE, a summary of the function results are printed to standard output.}
}
\value{

}
\description{
This function adds a simply adds a progress bar to \code{\link[dada2]{collapseNoMismatch}}.
}
\seealso{
\code{\link[dada2]{collapseNoMismatch}}
}
