% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_wrappers.R
\name{combineReadTrackingTables16S}
\alias{combineReadTrackingTables16S}
\title{Combine 16S Read-Tracking Tables}
\usage{
combineReadTrackingTables16S(
  trim_table,
  filter_table,
  dada_table,
  out_file = NULL
)
}
\arguments{
\item{trim_table}{Read-tracking table output by \code{\link{trimPrimers16S}}.}

\item{filter_table}{Read-tracking table output by \code{\link{qualityFilter16S}}.}

\item{dada_table}{Read-tracking table output by \code{\link{runDada16S}}.}

\item{out_file}{File location where a copy of the combined read-tracking table will be saved as a csv file. By default (NULL), this is \code{\link{NEONMICROBE_DIR_TRACKREADS}}/16S/track_16s_\link{timestamp}.csv. If no copy should be saved, set to FALSE.}
}
\value{
No value is returned.
}
\description{
Combine 16S Read-Tracking Tables
}
