% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combineReadTrackingTablesITS}
\alias{combineReadTrackingTablesITS}
\title{Combine ITS Read-Tracking Tables}
\usage{
combineReadTrackingTablesITS(prefilter_table, filter_table, dada_table)
}
\arguments{
\item{filter_table}{Read-tracking table output by \code{\link{qualityFilterITS}}.}

\item{dada_table}{Read-tracking table output by \code{\link{runDadaITS}}.}

\item{trim_table}{Read-tracking table output by \code{\link{trimPrimersITS}}.}
}
\value{
No value is returned.
}
\description{
Convenience function for combining the read-tracking tables
across the trimming, filtering, and DADA steps of the ITS
processing pipeline.
}
