% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downloadRawSequenceData}
\alias{downloadRawSequenceData}
\title{Download NEON Marker Gene Sequencing Raw Data}
\usage{
downloadRawSequenceData(
  metadata,
  outdir = PRESET_OUTDIR_SEQUENCE,
  verbose = FALSE
)
}
\arguments{
\item{metadata}{The output of downloadSequenceMetadata(). Must be provided as either the data.frame returned by downloadSequenceMetadata() or as a filepath to the csv file produced by downloadSequenceMetadata() when outdir is provided.}

\item{outdir}{Location where output files are saved. Defaults to PRESET_OUTDIR_SEQUENCE in params.R.}

\item{verbose}{If TRUE, prints status messages and progress bars associated with file downloads.}
}
\value{
Returns (invisibly) an integer code: 0 indicates success of downloads and a non-zero integer indicates failure. See the help page for \code{\link[utils]{download.file}} for more details.
}
\description{
Downloads NEON raw sequence data files to the specified filepath, by referencing
URLs in the metadata output from \code{\link{downloadSequenceMetadata}}.
}
