% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downloadRawSoilData}
\alias{downloadRawSoilData}
\title{Download NEON Soil Data Associated with Marker Gene Sequencing Data}
\usage{
downloadRawSoilData(
  sites = PRESET_SITES,
  startYrMo = PRESET_START_YR_MO,
  endYrMo = PRESET_END_YR_MO,
  outdir = PRESET_OUTDIR_SOIL,
  checkFileSize = PRESET_CHECK_FILE_SIZE,
  return_data = PRESET_RETURN_DATA,
  overwrite = FALSE
)
}
\arguments{
\item{sites}{Either the string 'all', meaning all available sites, or a character vector of 4-letter NEON site codes, e.g. c('ONAQ','RMNP'). Defaults to PRESET_SITES parameter in params.R.}

\item{startYrMo, endYrMo}{Either NA, meaning all available dates, or a character vector in the form YYYY-MM, e.g. 2017-01. Defaults to PRESET_START_YR_MO in params.R.}

\item{outdir}{Location where output files are saved. Defaults to PRESET_OUTDIR_SEQMETA in params.R.}

\item{checkFileSize}{TRUE or FALSE. Whether the user should be told the total file size before downloading. Defaults to PRESET_CHECK_FILE_SIZE in params.R.}

\item{return_data}{Whether to return metadata. If FALSE, simply downloads data to outdir and returns no value. Defaults to PRESET_RETURN_DATA.}

\item{overwrite}{TRUE or FALSE. If there is a previous download of this metadata in outdir, whether to overwrite that download.}
}
\value{
If return_data==TRUE, returns a dataframe consisting of joined soil data records from DP1.10078 ("Soil chemical properties (Distributed periodic)") and DP1.10086 ("Soil physical properties (Distributed periodic)"). Otherwise, no value is returned.
}
\description{
Downloads the following data products to the specified filepath:
(1) DP1.10078: "Soil chemical properties (Distributed periodic)";
(2) DP1.10086: "Soil physical properties (Distributed periodic)".
This function uses \code{\link[neonUtilities]{zipsByProduct}} to conduct the downloads.
}
