% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downloadSequenceMetadata}
\alias{downloadSequenceMetadata}
\title{Download NEON Marker Gene Sequencing Metadata}
\usage{
downloadSequenceMetadata(
  sites = PRESET_SITES,
  startYrMo = PRESET_START_YR_MO,
  endYrMo = PRESET_END_YR_MO,
  outdir = PRESET_OUTDIR_SEQMETA,
  checkFileSize = PRESET_CHECK_FILE_SIZE,
  return_data = PRESET_RETURN_DATA,
  target_genes = TARGET_GENE,
  sequencing_runs = SEQUENCING_RUNS,
  overwrite = FALSE
)
}
\arguments{
\item{sites}{Either the string 'all', meaning all available sites, or a character vector of 4-letter NEON site codes, e.g. c('ONAQ','RMNP'). Defaults to PRESET_SITES parameter in params.R.}

\item{startYrMo, endYrMo}{Either NA, meaning all available dates, or a character vector in the form YYYY-MM, e.g. 2017-01. Defaults to PRESET_START_YR_MO in params.R.}

\item{outdir}{Location where output files are saved. Defaults to PRESET_OUTDIR_SEQMETA in params.R.}

\item{checkFileSize}{TRUE or FALSE. Whether the user should be told the total file size before downloading. Defaults to PRESET_CHECK_FILE_SIZE in params.R.}

\item{return_data}{Whether to return part of the metadata: the table NEON table "mmg_soilRawDataFiles.csv" within NEON data product DP1.10108. If FALSE, simply downloads data to outdir and returns no value. Defaults to PRESET_RETURN_DATA.}

\item{target_genes}{'ITS', '16S', or 'all'. Defaults to TARGET_GENE in params.R.}

\item{sequencing_runs}{Either the string 'all', meaning all available sequencing runs, or a character vector of NEON sequencing run IDs, e.g. c('C25G9', 'B69PP'). Defaults to SEQUENCING_RUNS parameter in params.R.}

\item{overwrite}{TRUE or FALSE. If there is a previous download of this metadata in outdir, whether to overwrite that download.}
}
\value{
If return_data==TRUE, returns the mmg_soilRawDataFiles table from NEON.DP1.10108 ("Soil microbe marker gene sequences"). Otherwise, no value is returned.
}
\description{
Downloads NEON data product DP1.10108 ("Soil microbe marker gene sequences"),
which contains information on how to access raw sequence data. This function
uses \code{\link[neonUtilities]{zipsByProduct}} to conduct the downloads.
}
