% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downloadSequenceMetadataRev}
\alias{downloadSequenceMetadataRev}
\title{Download Sequence Metadata}
\usage{
downloadSequenceMetadataRev(
  sites = "all",
  startYrMo,
  endYrMo,
  targetGene = "",
  sequencingRuns = "",
  dpID = "DP1.10108.001",
  dir = ""
)
}
\arguments{
\item{sites}{Either the string 'all', meaning all available sites, or a character vector of 4-letter NEON site codes, e.g. c('ONAQ','RMNP'). Defaults to PRESET_SITES parameter in params.R.}

\item{startYrMo, endYrMo}{Either NA, meaning all available dates, or a character vector in the form YYYY-MM, e.g. 2017-01.}

\item{targetGene}{'16S' or 'ITS'.}

\item{sequencingRuns}{Either the string 'all', meaning all available sequencing runs, or a character vector of NEON sequencing run IDs, e.g. c('C25G9', 'B69PP').}

\item{dpID}{NEON data product of interest. Default is soil marker gene sequences, and currently code only works for this dpID.}

\item{dir}{(Optional) If a local copy of the filtered metadata is desired, provide path to output directory.}
}
\value{
Data frame containing joined records from across the NEON soil marker gene sequence metadata, subsetted according to function arguments.
}
\description{
Loads soil marker gene sequencing metadata for specified target gene, site(s) and date(s),
with an option to download output by providing a valid output directory.
}
\details{
Function by Lee F. Stanish and Clara Qin (2020). Currently available for testing only.
}
\examples{
\dontrun{
meta <- downloadSequenceMetadataRev('all', '2015-01', '2016-01', '16S') # metadata is not saved to local directory
meta <- downloadSequenceMetadataRev('all', '2015-01', '2016-01', '16S', dir='./data/') # metadata is saved to local directory
}
}
