% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neon_data.R
\name{downloadSoilData}
\alias{downloadSoilData}
\title{Download NEON Soil Data Associated with Marker Gene Sequencing Data}
\usage{
downloadSoilData(
  sites = "all",
  startYrMo = NA,
  endYrMo = NA,
  dpID = c("DP1.10086.001"),
  outDir = NEONMICROBE_DIR_SOIL(),
  rmSamplingImpractical = TRUE,
  rmNTransBouts = TRUE,
  rmFailedCNDataQF = TRUE
)
}
\arguments{
\item{sites}{Either the string 'all' (default), meaning all available sites, or a character vector of 4-letter NEON site codes, e.g. c('ONAQ','RMNP'). Defaults to PRESET_SITES parameter in params.R.}

\item{startYrMo, endYrMo}{Either NA (default), meaning all available dates, or a character vector in the form YYYY-MM, e.g. 2017-01. Defaults to PRESET_START_YR_MO in params.R.}

\item{dpID}{NEON data product(s) of interest. Default is DP1.10086.001 ("Soil physical and chemical properties, periodic").}

\item{outDir}{Default NEONMICROBE_DIR_SOIL() If a local copy of the filtered metadata is desired, provide path to output directory.}

\item{rmSamplingImpractical}{Default TRUE. Whether to remove soil data records when sampling did not actually occur.}

\item{rmNTransBouts}{Default TRUE. Whether to remove soil data records from bouts to collect N-transformation incubation tubes. These are only useful for calculating N-transformation rates and aren't associated with microbial data.}

\item{rmFailedCNDataQF}{Default TRUE. Whether to remove soil data records where cnPercentQF indicates failure. While other QF fields exist and are simply passed to output, this particular check may be desirable because this function later aggregates nitrogenPercent and organicCPercent values for cnSampleIDs with analytical replicates.}
}
\value{
If return_data==TRUE, returns a dataframe consisting of joined soil data records from DP1.10086 ("Soil physical and chemical properties, periodic"). Otherwise, no value is returned.
}
\description{
Downloads the following NEON data product:
\itemize{
\item DP1.10086.001: "Soil physical and chemical properties, periodic", tables sls_soilCoreCollection, sls_soilMoisture, sls_soilpH, and sls_soilChemistry.
This function uses \code{\link[neonUtilities]{loadByProduct}} to conduct the downloads.
}
}
