% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_management.R
\name{getBatch}
\alias{getBatch}
\title{Get Current Batch ID}
\usage{
getBatch(verbose = FALSE)
}
\arguments{
\item{verbose}{Default FALSE. If TRUE, returns information about the current processing batch.}
}
\value{
Character. Unique ID of the current processing batch.
}
\description{
Get the unique ID of the current processing batch.
}
\examples{
getBatch() # "xyz"
getBatch(verbose = TRUE) # "xyz", and table containing info about batch "xyz"
}
