% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_management.R
\name{getBatchParam}
\alias{getBatchParam}
\title{Get Processing Batch Parameter(s)}
\usage{
getBatchParam(params = NULL)
}
\arguments{
\item{params}{Parameters to get. By default (NULL), returns all parameters.}
}
\value{
Named list of parameters if multiple parameters are retrieved; else, single element (can be char, int, double).
}
\description{
Gets parameter(s) associated with the currently loaded processing batch. Based on \code{\link[dada2]{getDadaOpt}}.
}
\seealso{
\code{\link{setBatchParam}}
}
