% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getN}
\alias{getN}
\title{Get Number of Reads}
\usage{
getN(x)
}
\arguments{
\item{x}{A dada-class object.}
}
\value{
Integer; total number of reads in the dada-class object.
}
\description{
Gets the total number of reads in a dada-class object.
Primarily used for tracking reads throughout the pipeline.
}
