% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSampleName}
\alias{getSampleName}
\title{Get Sample Name from Fastq Filename (DEPRECATED)}
\usage{
getSampleName(
  fn,
  post_samplename_pattern1 = "_R1.*\\\\.fastq",
  post_samplename_pattern2 = "_R2.*\\\\.fastq"
)
}
\arguments{
\item{fn}{Fastq filename(s).}

\item{post_samplename_pattern1, post_samplename_pattern2}{Default "_R1.\emph{\\.fastq" and "_R2.}\\.fastq". Regular expressions denoting the ends of the R1 and R2 filenames, respectively, which will be trimmed away, leaving identical names between corresponding R1 and R2 files.}
}
\value{
Character vector; sample names of the input fastq filenames.
}
\description{
Deprecated in favor of code that reads dnaSampleID from sequence metadata.
}
\details{
Extracts the sample names from fastq filenames, given regular expressions
denoting the different endings of R1 and R2 filenames.
}
