% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getTruncationLength}
\alias{getTruncationLength}
\title{Get Truncation Length}
\usage{
getTruncationLength(fl, qscore = 30, n = 5e+05, verbose = TRUE)
}
\arguments{
\item{fl}{Full name(s) of fastq file(s).}

\item{qscore}{Default 30. Mean quality score threshold at which to truncate the remainder of the read.}

\item{n}{Default 5e+05. The number of reads to sample when processing fastq files.}

\item{verbose}{Default TRUE. Whether to return message regarding truncation length for each file. Includes warning messages.}
}
\value{
Integer vector of truncation lengths to use for the set of fastq files.
}
\description{
Decides on truncation length for trimmed reads based on quality score means. Default cutoff is a score of 30. Warns you if the beginning (first 10) bases are low-quality, but returns the first low-quality base after the 10th base. If no bases are below score, returns the last base. The truncation lengths can be aggregated (e.g. minimum) and used as an argument in \code{\link[dada2]{filterAndTrim}}.
}
\examples{
\dontrun{
trunc_len_fwd <- getTruncationLength(c("sample1_R1.fastq", "sample2_R1.fastq"))
trunc_len_rev <- getTruncationLength(c("sample1_R2.fastq", "sample2_R2.fastq"))
}
}
