% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_management.R
\name{listBatches}
\alias{listBatches}
\title{List Processing Batches}
\usage{
listBatches(batches_dir = NULL)
}
\arguments{
\item{batches_dir}{Directory where batch-specific directories are saved. By default (NULL), looks in NEONMICROBE_DIR_BATCHES().}
}
\value{

}
\description{
Lists all processing batches available in the specified batches directory, along with basic information about them.
}
