% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_management.R
\name{loadParams}
\alias{loadParams}
\title{Load Processing Parameters from File}
\usage{
loadParams(params_file, env = NULL, verbose = TRUE)
}
\arguments{
\item{params_file}{Filepath to R script containing name-value assignments (e.g. x = 1) that will be used to create an environment.}

\item{env}{Environment in which to load the parameters. By default (NULL), searches for current processing batch environment. If found, loads parameters there.}

\item{verbose}{Default TRUE. Whether to print warning message if a parameter has already been assigned in a processing batch.}
}
\value{

}
\description{
Loads neonMicrobe processing parameters from a specified file to a specified environment,
usually associated with a processing batch.
}
\examples{
\dontrun{
e <- new.env()
ls(e)  # character(0)
loadParams("./R/test_params.R", env=e)
ls(e)  # "x" "y"
get("x", e)  # "hello there"
get("y", e)  # "goodbye now"
loadParams("./R/test_params.R", env=e) # warning
}
}
