% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{makeDataDirectories}
\alias{makeDataDirectories}
\title{Make Data Directories}
\usage{
makeDataDirectories(check_location = TRUE)
}
\arguments{
\item{check_location}{Default TRUE. Whether to prompt the user to confirm that the data directory structure should be created in the current working directory.}
}
\value{
No value is returned.
}
\description{
Creates the directories necessary for storing output data, rooted in the specified directory.
}
\examples{
\dontrun{
makeDataDirectories()
}
}
\seealso{
\code{\link{setBaseDirectory}}
}
