% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{makeOutputDirectories}
\alias{makeOutputDirectories}
\title{Make Output Directories}
\usage{
makeOutputDirectories(
  seq_dir = PRESET_OUTDIR_SEQUENCE,
  seqmeta_dir = PRESET_OUTDIR_SEQMETA,
  soil_dir = PRESET_OUTDIR_SOIL,
  taxref_dir = PRESET_OUTDIR_TAXREF,
  outputs_dir = PRESET_OUTDIR_OUTPUTS
)
}
\arguments{
\item{seq_dir, seqmeta_dir, soil_dir, taxref_dir, outputs_dir}{Absolute paths to directory where each type of data will be written (sequence data, sequence metadata, soil data, taxonomic reference data, processing outputs). Defaults to the following preset parameters, respectively: PRESET_OUTDIR_SEQUENCE, PRESET_OUTDIR_SEQMETA, PRESET_OUTDIR_SOIL, PRESET_OUTDIR_TAXREF, PRESET_OUTDIR_OUTPUTS.}
}
\value{
No value is returned.
}
\description{
Make Output Directories
}
\examples{
\dontrun{
makeOutputDirectories()
}
}
