% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matchFastqToMetadata}
\alias{matchFastqToMetadata}
\title{Match Fastq Files to Metadata}
\usage{
matchFastqToMetadata(fn, meta, verbose = TRUE)
}
\arguments{
\item{fn}{Full name(s) of fastq file(s).}

\item{meta}{Metadata downloaded using \code{\link{downloadSequenceMetadata}} that corresponds to the fastq files.}

\item{verbose}{(Optional) Default TRUE. Whether to print warning messages when metadata does not contain records corresponding to all provided fastq files.}
}
\value{
Data frame, in which the first column contains the fastq file names in the order received, and the remaining columns contain corresponding metadata records. Has as many rows as the number of input filenames.
}
\description{
Helper function to match fastq file names to sequence metadata via
the sequence metadata table's "rawDataFileName" column. Tolerant to
filenames with appended run ID prefix (e.g. "runB69RN_...") and
agnostic to .gz compression.
}
\details{
Used in \code{\link{getPairedFastqFiles}} and \code{\link{removeUnpairedFastqFiles}}.
}
\examples{
\dontrun{
meta_ext <- matchFastqToMetadata(c("sample1_R1.fastq", "sample1_R2.fastq"), )
}
}
