% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirs.R
\name{NEONMICROBE_DIR_BASE}
\alias{NEONMICROBE_DIR_BASE}
\title{Dynamic Directory Name for Base Directory}
\usage{
NEONMICROBE_DIR_BASE()
}
\value{
Directory path (character).
}
\description{
Dynamic Directory Name for Base Directory
}
