% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirs.R
\name{NEONMICROBE_DIR_BATCHES}
\alias{NEONMICROBE_DIR_BATCHES}
\title{Dynamic Directory Name for Processing Batch Outputs}
\usage{
NEONMICROBE_DIR_BATCHES()
}
\value{
Directory path (character).
}
\description{
For batch-specific outputs (when processing batch is set using \code{\link{setBatch}} or \code{\link{newBatch}})
}
