% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirs.R
\name{NEONMICROBE_DIR_MIDPROCESS}
\alias{NEONMICROBE_DIR_MIDPROCESS}
\title{Dynamic Directory Name for Mid-Processing Fastq Files}
\usage{
NEONMICROBE_DIR_MIDPROCESS()
}
\value{
Directory path (character).
}
\description{
For sequence files in the middle of being processed. Subject
to modification by batch processing parameters via their
modifications to \code{link{NEONMICROBE_DIR_OUTPUTS}}, of which
this is a subdirectory. Contains subdirs "16S", "ITS".
}
