% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirs.R
\name{NEONMICROBE_DIR_OUTPUTS}
\alias{NEONMICROBE_DIR_OUTPUTS}
\title{Dynamic Directory Name for Outputs}
\usage{
NEONMICROBE_DIR_OUTPUTS(dir = NULL)
}
\arguments{
\item{dir}{If not NULL, directory path to force this to return.}
}
\value{
Directory path (character).
}
\description{
For outputs (sequence table, taxonomy table, phyloseq object).
Contains mid_process and track_reads. Subject to modification
by batch processing parameters.
}
