% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirs.R
\name{NEONMICROBE_DIR_SEQMETA}
\alias{NEONMICROBE_DIR_SEQMETA}
\title{Dynamic Directory Name for Sequence Metadata}
\usage{
NEONMICROBE_DIR_SEQMETA()
}
\value{
Directory path (character).
}
\description{
For metadata associated with NEON soil microbe marker gene sequences (DP1.10108.001).
}
