% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirs.R
\name{NEONMICROBE_DIR_SEQUENCE}
\alias{NEONMICROBE_DIR_SEQUENCE}
\title{Dynamic Directory Name for Raw Sequence Data}
\usage{
NEONMICROBE_DIR_SEQUENCE()
}
\value{
Directory path (character).
}
\description{
For raw sequence data (fastq files) from NEON soil microbe marker gene sequences
(DP1.10108.001).
}
