% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirs.R
\name{NEONMICROBE_DIR_SOIL}
\alias{NEONMICROBE_DIR_SOIL}
\title{Dynamic Directory Name for Soil Data}
\usage{
NEONMICROBE_DIR_SOIL()
}
\value{
Directory path (character).
}
\description{
For NEON soil data DP1.10086.001:
"Soil physical and chemical properties, periodic",
tables sls_soilCoreCollection, sls_soilMoisture, sls_soilpH, and sls_soilChemistry.
}
