% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirs.R
\name{NEONMICROBE_DIR_TAXREF}
\alias{NEONMICROBE_DIR_TAXREF}
\title{Dynamic Directory Name for Taxonomic Reference Data}
\usage{
NEONMICROBE_DIR_TAXREF()
}
\value{
Directory path (character).
}
\description{
For taxonomic reference data, e.g. UNITE and SILVA databases.
Not strictly necessary for functioning, so long as user specifies
reference datasets in \code{\link[dada2]{assignTaxonomy}}.
}
