% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirs.R
\name{NEONMICROBE_DIR_TRACKREADS}
\alias{NEONMICROBE_DIR_TRACKREADS}
\title{Dynamic Directory Name for Read-Tracking Tables}
\usage{
NEONMICROBE_DIR_TRACKREADS()
}
\value{
Directory path (character).
}
\description{
For read-tracking tables. Subject to modification by batch
processing parameters via their modifications to \code{link{NEONMICROBE_DIR_OUTPUTS}},
of which this is a subdirectory. Contains subdirs "16S", "ITS".
}
