% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_management.R
\name{newBatch}
\alias{newBatch}
\title{Create New Processing Batch}
\usage{
newBatch(
  seq_meta_file,
  batch_id = NULL,
  batches_dir = file.path(PRESET_OUTDIR_OUTPUTS, "batches"),
  overwrite = FALSE,
  set_batch = TRUE
)
}
\arguments{
\item{seq_meta_file}{Character string. File path to the sequence metadata to associate with this processing batch. Once set, this cannot be changed except by overwriting this batch.}

\item{batch_id}{Character string. Unique ID to use for the new processing batch.}

\item{batches_dir}{Default file.path(PRESET_OUTDIR_OUTPUS, "batches"). Directory where batch-specific directories are saved.}

\item{overwrite}{Default FALSE. If processing batch already exists in the specified directory, whether to overwrite it.}

\item{set_batch}{Default TRUE. Whether to switch to the new processing batch after creating it.}
}
\value{
No value returned
}
\description{
Create New Processing Batch
}
\examples{
\dontrun{
newBatch("abc", "data/sequence_metadata/mmg_soilMetadata_ITS_2021-03-08134134.csv")
newBatch("xyz", "data/sequence_metadata/mmg_soilMetadata_ITS_2021-03-08134134.csv")
}
}
