% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neon_data.R
\name{organizeRawSequenceData}
\alias{organizeRawSequenceData}
\title{Organize Raw Sequence Data}
\usage{
organizeRawSequenceData(
  fn,
  metadata,
  outdir_sequence = PRESET_OUTDIR_SEQUENCE,
  verbose = TRUE
)
}
\arguments{
\item{fn}{Character vector of full names (including path) of raw sequence files. Can include tarballs.}

\item{metadata}{The output of downloadSequenceMetadata(). Must be provided as either the data.frame returned by downloadSequenceMetadata() or as a filepath to the csv file produced by downloadSequenceMetadata() when outDir is provided.}

\item{outdir_sequence}{Default PRESET_OUTDIR_SEQUENCE. Directory where raw sequence files can be found before reorganizing.}

\item{verbose}{If TRUE, prints message each time a file is reorganized.}
}
\value{
Character vector of the files (including files within tarballs) that were successfully reorganized. If no files were successfully reorganized, returns no value.
}
\description{
Moves raw sequence data into the correct subdirectory for the processing pipeline,
renames files to include sequencer run ID, and untars sequence data if necessary.
}
