% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseParamsFromRownames}
\alias{parseParamsFromRownames}
\title{Parse Parameter Values from Rownames}
\usage{
parseParamsFromRownames(
  df,
  param1,
  param2,
  keep_rownames = FALSE,
  keep_original_cols = TRUE
)
}
\arguments{
\item{df}{Data frame with rownames containing parameter values of the form \link{param_name1}\emph{\link{param_value1}}\link{param_name2}_\link{param_value2}.}

\item{param1, param2}{Parameter names embedded within the rownames of df.}

\item{keep_rownames}{(Default FALSE) Whether to keep the rownames of the original dataframe.}

\item{keep_original_cols}{(Default TRUE) Whether to keep the columns of the original dataframe.}
}
\value{
A data frame containing columns corresponding to the values of the parameters of interest, extracted from the rownames of the original data frame.
}
\description{
Parses parameter values from rownames. Particularly useful for sensitivity analyses, when rows have been constructed to contain information about the parameter values, e.g. "maxEE.R_8_truncLen.R_250".
}
\examples{
\dontrun{ trackReadsWithParamVals <- parseParamsFromRownames(trackReads, PARAM1, PARAM2) }
}
