% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotEEProfile}
\alias{plotEEProfile}
\title{Plot Expected Errors Profile}
\usage{
plotEEProfile(
  fn,
  n = 5e+05,
  aggregate = FALSE,
  logEE = TRUE,
  include_quality_profile = FALSE
)
}
\arguments{
\item{fn}{Full name(s) of input fastq file(s), including directory.}

\item{n}{(Optional) Default 5e+5. The number of reads to sample when processing fastq files; passed to \code{\link[ShortRead]{qa}}.}

\item{aggregate}{(Optiona) Default FALSE. If inputting multiple fastq files, whether to aggregate their expected errors into one summary plot. If "aggregate" and "include_quality_profile" are both TRUE, also aggregates the quality profiles into one summary plot.}

\item{logEE}{(Optional) Default TRUE. Whether to log10-transform the Y-axis of the expected error profile.}

\item{include_quality_profile}{(Optional) Default FALSE. Whether to include the quality profile in the return value (see below).}
}
\value{
If include_quality_profile = FALSE (default), returns a ggplot of the expected errors. If include_quality_profile = TRUE, returns a length of list two: (1) a ggplot of the expected errors, and (2) a ggplot of the quality profile(s).
}
\description{
Plots the cumulative expected errors of a fastq file or collection of fastq files over sequence positions. This function extends \code{\link[dada2]{plotQualityProfile}} and makes use of its code.
}
\examples{
\dontrun{
fnFs <- c("sample1_R1.fastq", "sample2_R1.fastq", "sample3_R1.fastq")
plotEEProfile(fnFs) # produces a multi-faceted ggplot
plotEEProfile(fnFs, aggregate=TRUE) # produces a single-faceted ggplot
p <- plotEEProfile(fnFs, aggregate=TRUE, include_quality_profile=TRUE) # produces a list containing two single-faceted ggplots:
plot(p[[1]]) # expected errors profile
plot(p[[2]]) # quality profile
}
}
