% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_wrappers.R
\name{qualityFilter16S2}
\alias{qualityFilter16S2}
\title{Filter 16S Sequences (with metadata)}
\usage{
qualityFilter16S2(
  fn,
  dir_out,
  meta,
  trunc_qscore = 23,
  multithread = MULTITHREAD,
  ...
)
}
\arguments{
\item{fn}{Full names of input fastq files, including directory. Files that do not exist will be ignored; however, if all files do not exist, this function will issue a warning. It is assumed that these are R1 (forward-read) files only.}

\item{dir_out}{Directory where filtered fastq files will be written.}

\item{meta}{Metadata downloaded using \code{\link{downloadSequenceMetadata}} that corresponds to the fastq files.}

\item{trunc_qscore}{Default 23. Quality score at which point to truncate each read, if truncLen is NULL.}

\item{multithread}{Default MULTITHREAD in params.R. Whether to use multithreading. Note that Windows does not support multithreading in this function because it uses mclapply, so this argument must be set to FALSE on Windows systems.}

\item{...}{Other arguments to be passed to \code{\link[dada2]{filterAndTrim}}, such as maxEE and truncLen. See documentation for more details.}
}
\value{
Two-column matrix displaying the number of reads in input vs. output for each file.
}
\description{
Applies a quality filter to 16S sequence fastq files via the \code{\link[dada2]{filterAndTrim}} function.
It is assumed that both forward- and reverse-read files are included.
}
