% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_wrappers.R
\name{qualityFilterITS}
\alias{qualityFilterITS}
\title{Filter ITS Sequences}
\usage{
qualityFilterITS(
  fn,
  dir_in,
  dir_out,
  multithread = MULTITHREAD,
  post_samplename_pattern1 = "_R1.*\\\\.fastq",
  post_samplename_pattern2 = "_R2.*\\\\.fastq",
  ...
)
}
\arguments{
\item{fn}{Names of input fastq files, excluding directory path which is specified by dir_in. Files that do not exist will be ignored; however, if all files do not exist, this function will issue a warning. It is assumed that these are R1 (forward-read) files only.}

\item{dir_in}{Directory containing input fastq files.}

\item{dir_out}{Path to output directory where filtered fastq files will be written.}

\item{multithread}{Default MULTITHREAD in params.R. Whether to use multithreading. Note that Windows does not support multithreading in this function because it uses mclapply, so this argument must be set to FALSE on Windows systems.}

\item{post_samplename_pattern1, post_samplename_pattern2}{(Optional) Character pattern within the filename which immediately follows the end of the sample name. Defaults to "_R(1|2).*\\.fastq", as NEON fastq files typically consist of a sample name followed by "_R1.fastq" or "_R2.fastq", etc.}

\item{...}{Other arguments to be passed to \code{\link[dada2]{filterAndTrim}}, such as maxEE. See documentation for more details.}
}
\value{
Two-column matrix displaying the number of reads in input vs. output for each file.
}
\description{
Applies a quality filter to ITS sequence fastq files via the \code{\link[dada2]{filterAndTrim}} function.
Currently only supports filtering forward-read sequences.
}
