% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readSequenceMetadata}
\alias{readSequenceMetadata}
\title{Read Sequence Metadata into Function}
\usage{
readSequenceMetadata(metadata)
}
\arguments{
\item{metadata}{Either a data.frame returned from \code{\link{downloadSequenceMetadata}} or the filepath to a local csv copy of the output from \code{\link{downloadSequenceMetadata}}.}
}
\value{
data.frame of the metadata.
}
\description{
Helper function to accept two formats of metadata into various functions in this package: data.frame and filepath to a csv
}
