% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_unmatched_files}
\alias{remove_unmatched_files}
\title{Remove Unmatched Fastq Files}
\usage{
remove_unmatched_files(fnFs, fnRs, post_samplename_pattern = "_R")
}
\arguments{
\item{fnFs}{Full name(s) of fastq file(s) containing forward-read sequence data.}

\item{fnRs}{Full name(s) of fastq file(s) containing reverse-read sequence data.}

\item{post_samplename_pattern}{(Optional) Character pattern within the filename which immediately follows the end of the sample name. Defaults to "_R", as NEON fastq files typically consist of a sample name followed by "_R1.fastq" or "_R2.fastq", etc.}
}
\value{
List of length 2. The first element is a vector of forward-read files that have reverse-read counterparts; the second element is a vector of reverse-read files that have forward-read counterparts.
}
\description{
Removes any forward-read files that do not have reverse-read counterparts, and vise versa.
This function is necessary because dada2::filterAndTrim() will throw an error
if the forward-read files and the reverse-read files are mismatched.
}
\examples{
matched_fn <- remove_unmatched_files(c("sample1_R1.fastq", "sample2_R1.fastq"), c("sample1_R2.fastq", "sample2_R2.fastq", "sample3_R2.fastq"))
fnFs <- matched_fn[[1]]
fnRs <- matched_fn[[2]]
}
