% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeUnpairedFastqFiles}
\alias{removeUnpairedFastqFiles}
\title{Remove Unpaired Fastq Files}
\usage{
removeUnpairedFastqFiles(fnFs, fnRs, meta, value = TRUE, verbose = TRUE)
}
\arguments{
\item{fnFs, fnRs}{Full name(s) of fastq file(s) corresponding to R1 (forward) reads and R2 (reverse) reads, respectively. Tolerant to filenames with appended run ID prefix (e.g. "runB69RN_...") and agnostic to .gz compression.}

\item{meta}{Metadata downloaded using \code{\link{downloadSequenceMetadata}} that corresponds to the fastq files.}

\item{value}{(Optional) Default TRUE; returns file names. If FALSE, returns indices.}

\item{verbose}{(Optional) Default TRUE. Whether to print warning messages when metadata does not contain records corresponding to all provided fastq files.}
}
\value{
List of length 2; first element contains matched R1 (forward) file names, and second element contains corresponding R2 (reverse) file names.
}
\description{
Given a vector of fastq file names and corresponding metadata, outputs
a list containing matching R1 and R2 files in the same order.
Useful for providing files to functions that utilize
\code{\link[dada2]{filterAndTrim}}, such as \code{\link{trimPrimer16S}}
and \code{\link{runDadaITS}}.
}
\details{
See related function: \code{\link{getPairedFastqFiles}}.
}
\examples{
\dontrun{
#' matched_fn <- removeUnpairedFastqFiles(c("sample1_R1.fastq", "sample3_R1.fastq"), c("sample1_R2.fastq", "sample2_R2.fastq", "sample3_R2.fastq"), meta)
fnFs <- matched_fn[[1]] # "sample1_R1.fastq" "sample3_R1.fastq"
fnRs <- matched_fn[[2]] # "sample1_R2.fastq" "sample3_R2.fastq"
}
}
