% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{runDada16S}
\alias{runDada16S}
\title{Run Dada on Paired-End 16S Sequences}
\usage{
runDada16S(PATH_FILTERED, MULTITHREAD, VERBOSE = FALSE, seed = NULL)
}
\arguments{
\item{PATH_FILTERED}{Path to directory containing input fastq files.}

\item{MULTITHREAD}{Whether to use multithreading.}

\item{VERBOSE}{Default FALSE. Whether to print messages regarding the dimensions of the resulting sequence table and the distribution of sequence lengths.}

\item{seed}{(Optional) Integer to use as random seed for reproducibility.}
}
\value{
A list of three elements. \strong{seqtab} is the sequence table before removing chimeras, \strong{seqtab.nochim} is the sequence table after removing chimeras, and \strong{track} is a data frame displaying the number of reads remaining for each sample at various points throughout the processing pipeline.
}
\description{
Runs the Dada algorithm to infer sample composition from paired-end 16S fastq files.
This implementation is based on Ben Callahan's vignette at \url{https://benjjneb.github.io/dada2/bigdata_paired.html}.
}
\examples{
\dontrun{
seqtab.list <- runDada16S('./seq/filtered/', TRUE, TRUE, 1010100)
}
}
