% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_wrappers.R
\name{runDada16S2}
\alias{runDada16S2}
\title{Run Dada on Paired-End 16S Sequences (with metadata)}
\usage{
runDada16S2(
  fn,
  meta,
  out_seqtab = "",
  out_track = "",
  remove_chimeras = TRUE,
  multithread = MULTITHREAD,
  verbose = FALSE,
  seed = NULL,
  nbases = 1e+07
)
}
\arguments{
\item{fn}{Full names of input fastq files, including directory. Files that do not exist will be ignored; however, if all files do not exist, this function will issue a warning.}

\item{meta}{Metadata downloaded using \code{\link{downloadSequenceMetadata}} that corresponds to the fastq files.}

\item{out_seqtab, out_track}{(Optional) File locations where sequence table and read-tracking table (respectively) will be saved as csv files. If blank (default), will not save to file.}

\item{remove_chimeras}{(Optional) Default TRUE. Whether to remove chimeras from the sequence table. Currently only supports removal using \code{\link[dada2]{removeBimeraDenovo}} with the consensus method.}

\item{multithread}{Default MULTITHREAD in params.R. Whether to use multithreading.}

\item{verbose}{Default FALSE. Whether to print messages regarding the samples being processed, dimensions of the resulting sequence table, and the distribution of sequence lengths.}

\item{seed}{(Optional) Integer to use as random seed for reproducibility.}

\item{nbases}{(Optional) Number of bases to use for learning errors. Default 1e7.}
}
\value{
A list of two elements. \strong{seqtab} is the sequence table, with chimeras removed if remove_chimeras == TRUE, and \strong{track} is a data frame displaying the number of reads remaining for each sample at various points throughout the processing pipeline.
}
\description{
Runs the Dada algorithm to infer sample composition from paired-end 16S fastq files.
This implementation is based on Ben Callahan's vignette at \url{https://benjjneb.github.io/dada2/bigdata_paired.html}.
}
\examples{
\dontrun{
seqtab.list <- runDada16S(c("sample1_R1.fastq", "sample1_R2.fastq", "sample2_R1.fastq", "sample2_R2.fastq"), meta, seed=1010100)
}
}
