% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_wrappers.R
\name{runDadaITS}
\alias{runDadaITS}
\title{Run Dada on R1 ITS sequences}
\usage{
runDadaITS(
  fn,
  dir_in,
  multithread = MULTITHREAD,
  verbose = FALSE,
  seed = NULL,
  post_samplename_pattern1 = "_R1.*\\\\.fastq",
  post_samplename_pattern2 = "_R2.*\\\\.fastq",
  nbases = 1e+07
)
}
\arguments{
\item{fn}{Names of input fastq files, excluding directory path which is specified by dir_in. Files that do not exist will be ignored; however, if all files do not exist, this function will issue a warning. It is assumed that these are R1 (forward-read) files only.}

\item{dir_in}{Directory containing input fastq files.}

\item{multithread}{Default MULTITHREAD in params.R. Whether to use multithreading.}

\item{verbose}{Default FALSE. Whether to print messages regarding the dereplication step, the denoising step, and the dimensions of the resulting sequence table and the distribution of sequence lengths.}

\item{seed}{(Optional) Integer to use as random seed for reproducibility.}

\item{post_samplename_pattern1, post_samplename_pattern2}{(Optional) Character pattern within the filename which immediately follows the end of the sample name. Defaults to "_R(1|2).*\\.fastq", as NEON fastq files typically consist of a sample name followed by "_R1.fastq" or "_R2.fastq", etc.}

\item{nbases}{(Optional) Number of bases to use for learning errors. Default 1e7.}
}
\value{
A list of three elements. \strong{seqtab} is the sequence table before removing chimeras, \strong{seqtab.nochim} is the sequence table after removing chimeras, and \strong{track} is a data frame displaying the number of reads remaining for each sample at various points throughout the processing pipeline.
}
\description{
Runs the Dada algorithm to infer sample composition from R1 ITS fastq files.
We recommend using only the R1 reads because attempting to merge with R2
reads may result in less accurate representations of the fungal community
composition (Pauvert et al., 2019).
This implementation is based on Ben Callahan's vignettes at \url{https://benjjneb.github.io/dada2/bigdata.html}
and \url{https://benjjneb.github.io/dada2/ITS_workflow.html}.
}
\examples{
\dontrun{
seqtab.list <- runDadaITS(c("sample1_R1.fastq", "sample1_R2.fastq", "sample2_R1.fastq", "sample2_R2.fastq"), './seq/filtered/', seed=1010100)
}
}
