% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_management.R
\name{saveParameters}
\alias{saveParameters}
\title{Save Processing Parameters to File}
\usage{
saveParameters(params_file = NULL, env = NULL)
}
\arguments{
\item{params_file}{(Optional) Filepath to R script in which the environment will be saved as text of name-value assignments (e.g. x = 1). By default, uses file path in the current batch directory called "params.R".}

\item{env}{Environment containing parameters to be saved. By default (NULL), searches for current processing batch environment. If found, saves parameters from there.}
}
\value{

}
\description{
Saves neonMicrobe processing parameters from a specified environment to a specified file,
usually associated with a processing batch.
}
