% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seqmeta_greatplains}
\alias{seqmeta_greatplains}
\title{16S Sequence Metadata for NEON Great Plains Sites}
\format{
A data frame with 172 rows and 86 variables.
For variable definitions, see NEON soil microbe
marker gene sequences data product:
\url{https://data.neonscience.org/data-products/DP1.10108.001}
}
\source{
downloadSequenceMetadata(startYrMo = "2017-07", endYrMo = "2017-07",
sites = c("KONZ", "CPER", "NOGP"),
targetGene = "16S", outDir=FALSE)
}
\usage{
seqmeta_greatplains
}
\description{
A subset of the NEON 16S sequence metadata for three
terrestrial NEON sites in the Great Plains: KONZ,
CPER, and NOGP. This sequence metadata contains
all records of 16S soil marker gene sequence data
collected from these sites in July 2017. This
sequence metadata has not been QCd using
\code{\link{qcMetadata}}.
}
\keyword{datasets}
