% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seqtab_greatplains}
\alias{seqtab_greatplains}
\title{16S Sequence Abundance Table for NEON Great Plains Sites}
\format{
A numeric matrix with 86 rows (samples) and
46,730 columns (ASVs). Values represent read abundances.
The table has been collapsed (100\% clustering) using
\code{\link[dada2]{collapseNoMismatch}}. Read abundances
are not standardized.
}
\source{
See vignette "Analyze NEON Great Plains 16S".
}
\usage{
seqtab_greatplains
}
\description{
A sequence abundance table generated from a
subset of the NEON 16S sequence metadata for three
terrestrial NEON sites in the Great Plains (KONZ,
CPER, and NOGP) collected during July 2017. This
sequence table has been collapsed using
\code{\link[dada2]{collapseNoMismatch}}.
}
\details{
This sequence abundance table was processed using maxEE=8
and truncLen=220 for forward and reverse reads. See the
full code used to generate this in the vignette
"Analyze NEON Great Plains 16S".
}
\keyword{datasets}
