% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setBaseDirectory}
\alias{setBaseDirectory}
\title{Set Base Directory for neonMicrobe Workspace}
\usage{
setBaseDirectory(dir = getwd())
}
\arguments{
\item{dir}{Directory to use as the base directory. Defaults to the working directory.}
}
\value{
No value is returned.
}
\description{
Sets the specified directory as the workspace for the neonMicrobe pipeline. This base directory persists until the end of the R session. If data directories have not yet been generated within the base directory, use \code{\link{makeDataDirectories}} to do so.
}
\seealso{
\code{\link{makeDataDirectories}}
}
