% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_management.R
\name{setBatch}
\alias{setBatch}
\title{Switch to Existing Processing Batch}
\usage{
setBatch(
  batch_id = NULL,
  batches_dir = file.path(NEONMICROBE_DIR_OUTPUTS(), "batches")
)
}
\arguments{
\item{batch_id}{Character string. Unique ID of an existing processing batch.}

\item{batches_dir}{Default file.path(NEONMICROBE_DIR_OUTPUTS(), "batches"). Directory where batch-specific directories are saved.}
}
\value{

}
\description{
Switch to Existing Processing Batch
}
