% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_management.R
\name{setBatchParameter}
\alias{setBatchParameter}
\title{Set Processing Batch Parameter(s)}
\usage{
setBatchParameter(...)
}
\arguments{
\item{...}{The processing parameters to set, along with their new value.}
}
\value{
No value is returned.
}
\description{
Sets parameter(s) associated with the currently loaded processing batch. Based on \code{\link[dada2]{setDadaOpt}}.
}
\details{
Quality filtering parameters, used as defaults in \code{\link[dada2]{filterAndTrim}}:

MAX_EE_FWD, MAX_EE_REV: max. allowable expected errors in forward/reverse reads that pass filter
MIN_LEN_FWD, MIN_LEN_REV: min. allowable length of reads that pass filter in forward/reverse reads
TRUNC_Q_FWD, TRUNC_Q_REV: base quality score after which to truncate the sequence in forward/reverse reads
TRUNC_LEN_FWD, TRUNC_LEN_REV: length of read after which to truncate the sequence in forward/reverse reads. Reads shorter than this value are discarded.

Parameters for DADA options, including heuristics: see \code{\link[dada2]{setDadaOpt}}
}
\seealso{
\code{\link[dada2]{filterAndTrim}}, \code{\link[dada2]{setDadaOpt}}
}
