% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soil_greatplains}
\alias{soil_greatplains}
\title{Soil Physical and Chemical Properties for NEON Great Plains Sites}
\format{
A numeric matrix with 86 rows (samples) and
46,730 columns (ASVs). Values represent read abundances.
The table has been collapsed (100\% clustering) using
\code{\link[dada2]{collapseNoMismatch}}. Read abundances
are not standardized.
}
\source{
See vignette "Analyze NEON Great Plains 16S".
}
\usage{
soil_greatplains
}
\description{
A data frame containing soil physical and chemical
properties data (DP1.10086.001: "Soil physical and chemical properties,
periodic", tables sls_soilCoreCollection, sls_soilMoisture, sls_soilpH,
and sls_soilChemistry) for three terrestrial NEON sites in the Great
Plains (KONZ, CPER, and NOGP) collected during July 2017.
}
\details{
This sequence abundance table was processed using maxEE=8
and truncLen=220 for forward and reverse reads. See the
full code used to generate this in the vignette
"Analyze NEON Great Plains 16S".
}
\keyword{datasets}
