% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trimPrimers16S}
\alias{trimPrimers16S}
\title{Trim Primers from 16S Sequences}
\usage{
trimPrimers16S(
  fnFs,
  fnRs,
  PATH_CUT,
  PRIMER_16S_FWD,
  PRIMER_16S_REV,
  MULTITHREAD
)
}
\arguments{
\item{fnFs}{Full name(s) of fastq file(s) containing forward-read sequence data.}

\item{fnRs}{Full name(s) of fastq file(s) containing reverse-read sequence data.}

\item{PATH_CUT}{Output directory. If it does not exist, it will be created.}

\item{PRIMER_16S_FWD}{DNA sequence of forward-read primer.}

\item{PRIMER_16S_REV}{DNA sequence of reverse-read primer.}

\item{MULTITHREAD}{Whether to use multithreading. Note that Windows does not support multithreading in this function because it uses mclapply, so this argument must be set to FALSE on Windows systems.}
}
\value{
Integer matrix denoting the number of reads remaining after primer-trimming for each input file.
}
\description{
Trims primers from 16S sequences using \code{\link[dada2]{filterAndTrim}}. This function
assumes that each read begins with its full primer sequence and operates
by truncating the beginning of each read by the length of its primer
sequence.
}
\examples{
\dontrun{
trimPrimers16S(c("sample1_R1.fastq", "sample2_R1.fastq"), c("sample1_R2.fastq", "sample2_R2.fastq"), , "CCTACGGGNBGCASCAG", "GACTACNVGGGTATCTAATCC", MULTITHREAD = TRUE)
}
}
