% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_wrappers.R
\name{trimPrimers16S2}
\alias{trimPrimers16S2}
\title{Trim Primers from 16S Sequences (with metadata)}
\usage{
trimPrimers16S2(
  fn,
  dir_out,
  meta,
  primer_16S_fwd = "CCTACGGGNBGCASCAG",
  primer_16S_rev = "GACTACNVGGGTATCTAATCC",
  multithread = MULTITHREAD
)
}
\arguments{
\item{fn}{Full names of input fastq files, including directory. Files that do not exist will be ignored; however, if all files do not exist, this function will issue a warning.}

\item{dir_out}{Directory where filtered fastq files will be written.}

\item{meta}{Metadata downloaded using \code{\link{downloadSequenceMetadata}} that corresponds to the fastq files.}

\item{primer_16S_fwd, primer_16S_rev}{DNA sequences of 16S forward and reverse primer, respectively}

\item{multithread}{Default MULTITHREAD in params.R. Whether to use multithreading. Note that Windows does not support multithreading in this function because it uses mclapply, so this argument must be set to FALSE on Windows systems.}
}
\value{
Integer matrix denoting the number of reads remaining after primer-trimming for each input file.
}
\description{
Trims primers from 16S sequences using \code{\link[dada2]{filterAndTrim}}. This function
assumes that each read begins with its full primer sequence and operates
by truncating the beginning of each read by the length of its primer
sequence.
}
\examples{
\dontrun{
trimPrimers16S(c("sample1_R1.fastq", "sample1_R2.fastq", "sample2_R1.fastq", "sample2_R2.fastq"), "path/to/output", meta, "CCTACGGGNBGCASCAG", "GACTACNVGGGTATCTAATCC", multithread = TRUE)
}
}
