% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_wrappers.R
\name{trimPrimersITS2}
\alias{trimPrimersITS2}
\title{Trim Primers from ITS Sequences (with metadata)}
\usage{
trimPrimersITS2(
  fn,
  dir_out,
  meta,
  cutadapt_path,
  primer_ITS_fwd = "CTTGGTCATTTAGAGGAAGTAA",
  primer_ITS_rev = "GCTGCGTTCTTCATCGATGC",
  very_verbose = FALSE,
  discard_untrimmed = FALSE
)
}
\arguments{
\item{fn}{Full names of input fastq files, including directory. Files that do not exist will be ignored; however, if all files do not exist, this function will issue a warning. It is assumed that these are R1 (forward-read) files only.}

\item{dir_out}{Directory where filtered fastq files will be written.}

\item{meta}{Metadata downloaded using \code{\link{downloadSequenceMetadata}} that corresponds to the fastq files.}

\item{cutadapt_path}{Default CUTADAPT_PATH in params.R. Path to cutadapt on your file system.}

\item{primer_ITS_fwd, primer_ITS_rev}{DNA sequence of the ITS forward and reverse primer, respectively.}

\item{very_verbose}{Default FALSE. Whether to print output from cutadapt.}

\item{discard_untrimmed}{Default FALSE. Whether to discard reads where a primer could not be found, leaving only those reads in which a primer has been trimmed.}
}
\value{
No value is returned.
}
\description{
Trims primers from ITS sequences using cutadapt. Cutadapt must be installed in order for this to work. Currently only supports forward-read (R1) sequences. If reverse reads are included, they will be ignored.
}
\examples{
\dontrun{
trimPrimersITS(c("sample1_ITS_R1.fastq", "sample2_ITS_R1.fastq"), "path/to/input", "path/to/output", "CTTGGTCATTTAGAGGAAGTAA")
}
}
