% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_wrappers.R
\name{validateIO}
\alias{validateIO}
\title{Validate Input/Output Arguments to DADA2 Wrappers}
\usage{
validateIO(
  fn,
  in_subdir,
  out_subdir,
  in_explicitdir,
  out_explicitdir,
  targetGene = c("16S", "ITS"),
  validate_input_only = FALSE
)
}
\arguments{
\item{fn}{Base names of input fastq files. If inputs are not base names (i.e. if they include directory paths), the directory paths will be removed. Files that do not exist will be ignored; however, if all files do not exist, this function will issue a warning.}

\item{in_subdir}{Subdirectory name from which to retrieve input fastq files. Enter "raw" for raw sequence files, or any other character string to specify a subdirectory within \code{\link{NEONMICROBE_DIR_MIDPROCESS}}/16S. To specify a directory outside \code{\link{NEONMICROBE_DIR_MIDPROCESS}}/16S, use the 'in_explicitdir' argument.}

\item{out_subdir}{Subdirectory name where output fastq files will be written. Enter any character string to specify a subdirectory within \code{\link{NEONMICROBE_DIR_MIDPROCESS}}/16S. If the directory does not exist, it will be created. To specify a directory outside \code{\link{NEONMICROBE_DIR_MIDPROCESS}}/16S, use the 'out_explicitdir' argument.}

\item{in_explicitdir, out_explicitdir}{Directory names to use instead of 'in_subdir' and 'out_subdir', if static directory names or directories outside of \code{\link{NEONMICROBE_DIR_MIDPROCESS}}/16S are desired. Not recommended for use within processing batches.}

\item{targetGene}{Either "16S" or "ITS".}
}
\value{
List.
\itemize{
\item fn: Full paths to input files.
\item fn_out: Full paths where each file will be written after processing.
\item in_dir: Full path to input directory.
\item out_dir: Full path to output directory.
}
}
\description{
Internal function for validating the input and output path arguments
to any of the DADA2 wrapper functions. Keeps only existing input
files. Output directory is modifiable by processing batches via
\code{\link{NEONMICROBE_DIR_OUTPUTS}}.
}
