% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{warn_already_downloaded}
\alias{warn_already_downloaded}
\title{Warn That Metadata Is Already Downloaded}
\usage{
warn_already_downloaded(PRNUM, outdir)
}
\arguments{
\item{PRNUM}{NEON data product ID, e.g. "10108".}

\item{outdir}{Directory where data product has been downloaded (not the stacked folder within it).}
}
\value{
No value is returned.
}
\description{
Issue a warning that metadata has already been downloaded, and indicates the sites and date
ranges for which it was downloaded. To be called within other functions.
}
