% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_sample_subset_params}
\alias{write_sample_subset_params}
\title{Write sample subsetting parameters}
\usage{
write_sample_subset_params(
  writeDir,
  sites,
  startYrMo,
  endYrMo,
  target_genes = "",
  sequencing_runs = ""
)
}
\arguments{
\item{writeDir}{Directory where the subsetting parameters should be written.}

\item{sites}{Vector containing the names of sites in the subset.}

\item{startYrMo, endYrMo}{Dates describing the time range of subset, formatted as "YYYY-MM".}

\item{target_genes}{Marker gene used for sequencing. Currently must be one of "16S", "ITS".}

\item{sequencing_runs}{Vector naming the sequencing runs in the subset.}
}
\value{
No value is returned.
}
\description{
Write site and date range parameters for subsetting samples. To be called within other functions.
}
\details{
This function writes the sample subsetting parameters to a file. In addition to
keeping a record, it is also meant to be used with \code{\link{warn_already_downloaded}}
to detect when a data product has already been downloaded, and if so, with what subsetting
parameters.
}
\examples{
\dontrun{
write_sample_subset_params(writeDir="./raw_data/", sites="all", startYrMo="2016-01", endYrMo="2019-01", target_genes="ITS", sequencing_runs=c("B69PP","C25G9"))
}
}
