% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_management.R
\name{checkArgsAgainstBatchParams}
\alias{checkArgsAgainstBatchParams}
\title{Check Function Arguments against Batch-Specific Parameters}
\usage{
checkArgsAgainstBatchParams(priority = c("batch", "arguments"), ...)
}
\arguments{
\item{priority}{Must be "batch" or "arguments". Whether to defer to batch-specific processing parameters or function arguments when they conflict.}

\item{...}{Arguments that have values defined in parameters, set equal to whatever value they have in the parent function, e.g. \code{MAX_EE_FWD = maxEE} These arguments will be the ones checked against the batch parameters.}
}
\value{
No value is returned
}
\description{
If a processing batch is currently set, inserting this function into another function from
neonMicrobe will check the parent function's arguments against the parameters associated
with the current processing batch. If priority = "batch", the current batch's parameters would
take precedence.
}
\details{
For example, if priority = "batch", then rather than using quality filtering parameters defined on-the-fly,
the parent function would use the quality filtering parameters associated with the current batch.
}
