% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combineReadTrackingTables16S}
\alias{combineReadTrackingTables16S}
\title{Combine 16S Read-Tracking Tables}
\usage{
combineReadTrackingTables16S(trim_table, filter_table, dada_table)
}
\arguments{
\item{trim_table}{Read-tracking table output by \code{\link{trimPrimers16S}}.}

\item{filter_table}{Read-tracking table output by \code{\link{qualityFilter16S}}.}

\item{dada_table}{Read-tracking table output by \code{\link{runDada16S}}.}
}
\value{
No value is returned.
}
\description{
Convenience function for combining the read-tracking tables
across the trimming, filtering, and DADA steps of the 16S
processing pipeline.
}
