% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combineReadTrackingTablesITS}
\alias{combineReadTrackingTablesITS}
\title{Combine ITS Read-Tracking Tables}
\usage{
combineReadTrackingTablesITS(prefilter_table, filter_table, dada_table)
}
\arguments{
\item{filter_table}{Read-tracking table output by \code{\link{qualityFilterITS}}.}

\item{dada_table}{Read-tracking table output by \code{\link{runDadaITS}}.}

\item{trim_table}{Read-tracking table output by \code{\link{trimPrimersITS}}.}
}
\value{
An integer matrix with columns representing each step in the DADA2
ITS processing pipeline, and rows representing each input sample,
with values representing the read counts remaining at that step
for each sample.
}
\description{
Convenience function for combining the read-tracking tables
across the trimming, filtering, and DADA steps of the ITS
processing pipeline.
}
