% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neon_data.R
\name{downloadRawSequenceData}
\alias{downloadRawSequenceData}
\title{Download NEON Marker Gene Sequencing Raw Data}
\usage{
downloadRawSequenceData(
  metadata,
  outDir = NULL,
  overwrite = FALSE,
  ignore_tar_files = TRUE,
  checkSize = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{metadata}{The output of \code{\link{downloadSequenceMetadata}}. Must be provided as either the data.frame returned by \code{\link{downloadSequenceMetadata}} or as a filepath to the csv file produced by \code{\link{downloadSequenceMetadata}}.}

\item{outDir}{Location where output files are saved and sorted into target gene-specific subdirectories. By default (NULL), this is NEONMICROBE_DIR_SEQUENCE().}

\item{overwrite}{Default FALSE. If TRUE, overwrites existing files.}

\item{ignore_tar_files}{If TRUE (default), does not download tar files. Each tar file is a batch containing an entire sequence run of fastq files. The tar file structure will soon be deprecated.}

\item{verbose}{If TRUE, prints status messages associated with file downloads (in addition to default progress bars).}
}
\value{
Returns (invisibly) a list of integer codes: 0 indicates success of downloads and a non-zero integer indicates failure. See the help page for \code{\link[utils]{download.file}} for more details.
}
\description{
Downloads NEON raw sequence data files to the specified filepath, by referencing
URLs in the metadata output from \code{\link{downloadSequenceMetadata}}.
}
