% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neon_data.R
\name{downloadRawSequenceData}
\alias{downloadRawSequenceData}
\title{Download NEON Marker Gene Sequencing Raw Data}
\usage{
downloadRawSequenceData(
  metadata,
  outDir = PRESET_OUTDIR_SEQUENCE,
  ignore_tar_files = TRUE,
  checkSize = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{metadata}{The output of downloadSequenceMetadata(). Must be provided as either the data.frame returned by downloadSequenceMetadata() or as a filepath to the csv file produced by downloadSequenceMetadata() when outdir is provided.}

\item{ignore_tar_files}{If TRUE (default), does not download tar files. Each tar file is a batch containing an entire sequence run of fastq files. The tar file structure will soon be deprecated.}

\item{verbose}{If TRUE, prints status messages and progress bars associated with file downloads.}

\item{outdir}{Location where output files are saved. Defaults to PRESET_OUTDIR_SEQUENCE in params.R.}
}
\value{
Returns (invisibly) a list of integer codes: 0 indicates success of downloads and a non-zero integer indicates failure. See the help page for \code{\link[utils]{download.file}} for more details.
}
\description{
Downloads NEON raw sequence data files to the specified filepath, by referencing
URLs in the metadata output from \code{\link{downloadSequenceMetadata}}.
}
