% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neon_data.R
\name{downloadRawSoilData}
\alias{downloadRawSoilData}
\title{Alias for downloadSoilData()}
\usage{
downloadRawSoilData(
  sites = "all",
  startYrMo = NA,
  endYrMo = NA,
  dpID = c("DP1.10086.001"),
  outDir = NEONMICROBE_DIR_SOIL(),
  rmSamplingImpractical = TRUE,
  rmNTransBouts = TRUE,
  rmFailedCNDataQF = TRUE
)
}
\arguments{
\item{rmFailedCNDataQF}{}
}
\description{
This function name is deprecated.
}
\seealso{
\code{\link{downloadSoilData}}
}
