% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neon_data.R
\name{downloadSequenceMetadata}
\alias{downloadSequenceMetadata}
\title{Download Sequence Metadata}
\usage{
downloadSequenceMetadata(
  sites = "all",
  startYrMo = NA,
  endYrMo = NA,
  targetGene = "all",
  sequencingRuns = "",
  dpID = "DP1.10108.001",
  outDir = file.path(NEONMICROBE_DIR_SEQMETA(), "raw_sequence")
)
}
\arguments{
\item{sites}{Either the string 'all', meaning all available sites, or a character vector of 4-letter NEON site codes, e.g. c('ONAQ','RMNP'). Defaults to PRESET_SITES parameter in params.R.}

\item{startYrMo, endYrMo}{Either NA (default), meaning all available dates, or a character vector in the form YYYY-MM, e.g. 2017-01.}

\item{targetGene}{'16S' or 'ITS'.}

\item{sequencingRuns}{Either the string 'all', meaning all available sequencing runs, or a character vector of NEON sequencing run IDs, e.g. c('C25G9', 'B69PP').}

\item{dpID}{NEON data product of interest. Default is soil marker gene sequences, and currently code only works for marker genes data products.}

\item{outDir}{Default file.path(NEONMICROBE_DIR_SEQMETA(), "raw_sequence"). To save a copy of the metadata to the local file system, provide the path to output directory. If this behavior is not desired, set outDir=FALSE.}
}
\value{
Data frame containing joined records from across the NEON soil marker gene sequence metadata, subsetted according to function arguments.
}
\description{
Loads soil marker gene sequencing metadata for specified target gene, site(s) and date(s),
with an option to download output by providing a valid output directory. This function uses
\code{\link[neonUtilities]{loadByProduct}} to conduct the downloads.
}
\details{
Function by Lee F. Stanish and Clara Qin (2020). Currently available for testing only.
}
\examples{
\dontrun{
meta <- downloadSequenceMetadataRev('all', '2015-01', '2016-01', '16S', outDir=FALSE) # metadata is not saved to local directory
meta <- downloadSequenceMetadataRev('all', '2015-01', '2016-01', '16S', outDir='./data/') # metadata is saved to local directory
}
}
