% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_management.R
\name{listBatches}
\alias{listBatches}
\title{List Processing Batches}
\usage{
listBatches(batches_dir = file.path(NEONMICROBE_DIR_OUTPUTS, "batches"))
}
\arguments{
\item{batches_dir}{}
}
\value{

}
\description{
Lists all processing batches available in the specified batches directory, along with basic information about them.
}
