% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_management.R
\name{listBatches}
\alias{listBatches}
\title{List Processing Batches}
\usage{
listBatches(batches_dir = file.path(PRESET_OUTDIR_OUTPUTS, "batches"))
}
\arguments{
\item{batches_dir}{}
}
\value{

}
\description{
List Processing Batches
}
