% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadParams}
\alias{loadParams}
\title{(Re)Load Parameters from Files}
\usage{
loadParams(
  user_params = "params/user.R",
  system_params = "params/system.R",
  validity_tests = "params/.test_params.R"
)
}
\arguments{
\item{user_params}{(Optional) Location of the R script containing user-defined parameters.}

\item{system_params}{(Optional) Location of the R script containing system parameters.}

\item{validity_tests}{(Optional) Location of the R script containing parameter validity tests.}
}
\value{
No value is returned.
}
\description{
(Re)loads parameters from files in the 'params' directory to the global environment.
}
