% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qualityFilter16S}
\alias{qualityFilter16S}
\title{Filter 16S Sequences}
\usage{
qualityFilter16S(
  PATH_CUT,
  PATH_FILTERED,
  MULTITHREAD,
  MAX_EE_FWD,
  MAX_EE_REV,
  TRUNC.LENGTHS = NULL,
  trunc_qscore = 23,
  post_samplename_pattern = "_R"
)
}
\arguments{
\item{PATH_CUT}{Path to directory containing input fastq files.}

\item{PATH_FILTERED}{Path to output directory where filtered fastq files will be written.}

\item{MULTITHREAD}{Whether to use multithreading. Note that Windows does not support multithreading in this function because it uses mclapply, so this argument must be set to FALSE on Windows systems.}

\item{MAX_EE_FWD, MAX_EE_REV}{The maximum number of expected errors allowed in forward/reverse reads. Read-pairs with expected errors that exceed this threshold will be removed.}

\item{TRUNC.LENGTHS}{Default NULL. Single integer: truncation length to use across all files. Two-integer vector: truncation length to use for the forward-read and reverse-read files, respectively. If NULL (default), determines truncation length(s) based on \code{\link{getTruncationLength}} with a quality score threshold of trunc_qscore.}

\item{trunc_qscore}{Default 23. Quality score at which point to truncate each read, if TRUNC.LENGTHS is null.}

\item{post_samplename_pattern}{(Optional) Character pattern within the filename which immediately follows the end of the sample name. Defaults to "_R", as NEON fastq files typically consist of a sample name followed by "_R1.fastq" or "_R2.fastq", etc.}
}
\value{
Two-column matrix displaying the number of reads in input vs. output for each file.
}
\description{
Applies a quality filter to 16S sequence fastq files via the \code{\link[dada2]{filterAndTrim}} function.
It is assumed that both forward- and reverse-read files are included.
}
