% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_wrappers.R
\name{qualityFilterITS}
\alias{qualityFilterITS}
\title{Filter ITS Sequences (with metadata)}
\usage{
qualityFilterITS(fn, dir_out, meta, multithread = FALSE, ...)
}
\arguments{
\item{fn}{Full names of input fastq files, including directory. Files that do not exist will be ignored; however, if all files do not exist, this function will issue a warning. It is assumed that these are R1 (forward-read) files only.}

\item{dir_out}{Directory where filtered fastq files will be written.}

\item{meta}{The output of \code{\link{downloadSequenceMetadata}}. Must be provided as either the data.frame returned by \code{\link{downloadSequenceMetadata}} or as a filepath to the csv file produced by \code{\link{downloadSequenceMetadata}}.}

\item{multithread}{Default FALSE. Whether to use multithreading. Note that Windows does not support multithreading in this function because it uses mclapply, so this argument must be set to FALSE on Windows systems.}

\item{...}{Other arguments to be passed to \code{\link[dada2]{filterAndTrim}}, such as maxEE, truncLen, and minLen. By default, uses \code{\link[dada2]{filterAndTrim}}'s default values. See documentation for \code{\link[dada2]{filterAndTrim}} for more details.}
}
\value{
(Invisibly) Two-column matrix displaying the number of reads in input vs. output for each file.
}
\description{
Applies a quality filter to ITS sequence fastq files via the \code{\link[dada2]{filterAndTrim}} function.
Currently only supports filtering forward-read (R1) sequences. If reverse reads are included, they will be ignored.
}
