% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_wrappers.R
\name{runDada16S}
\alias{runDada16S}
\title{Run Dada on Paired-End 16S Sequences}
\usage{
runDada16S(
  fn,
  dir_in,
  multithread = FALSE,
  verbose = FALSE,
  seed = NULL,
  post_samplename_pattern1 = "_R1.*\\\\.fastq",
  post_samplename_pattern2 = "_R2.*\\\\.fastq",
  nbases = 1e+07
)
}
\arguments{
\item{fn}{Names of input fastq files, excluding directory path which is specified by dir_in. Files that do not exist will be ignored; however, if all files do not exist, this function will issue a warning.}

\item{dir_in}{Directory containing input fastq files.}

\item{multithread}{Default FALSE. Whether to use multithreading.}

\item{verbose}{Default FALSE. Whether to print messages regarding the dimensions of the resulting sequence table and the distribution of sequence lengths.}

\item{seed}{(Optional) Integer to use as random seed for reproducibility.}

\item{nbases}{(Optional) Number of bases to use for learning errors. Default 1e7.}
}
\value{
A list of three elements. \strong{seqtab} is the sequence table before removing chimeras, \strong{seqtab.nochim} is the sequence table after removing chimeras, and \strong{track} is a data frame displaying the number of reads remaining for each sample at various points throughout the processing pipeline.
}
\description{
Runs the Dada algorithm to infer sample composition from paired-end 16S fastq files.
This implementation is based on Ben Callahan's vignette at \url{https://benjjneb.github.io/dada2/bigdata_paired.html}.
}
\examples{
\dontrun{
seqtab.list <- runDada16S(c("sample1_R1.fastq", "sample1_R2.fastq", "sample2_R1.fastq", "sample2_R2.fastq"), seed=1010100)
}
}
