% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_wrappers.R
\name{runDada16S}
\alias{runDada16S}
\title{Run Dada on Paired-End 16S Sequences (with metadata)}
\usage{
runDada16S(
  fn,
  in_subdir,
  meta,
  out_seqtab = NULL,
  out_track = NULL,
  in_explicitdir = NULL,
  remove_chimeras = TRUE,
  multithread = FALSE,
  verbose = FALSE,
  seed = NULL,
  nbases = 1e+07
)
}
\arguments{
\item{fn}{Base names of input fastq files. If inputs are not base names (i.e. if they include directory paths), the directory paths will be removed. Files that do not exist will be ignored; however, if all files do not exist, this function will issue a warning.}

\item{in_subdir}{Subdirectory name from which to retrieve input fastq files. Enter "raw" for raw sequence files, or any other character string to specify a subdirectory within \code{\link{NEONMICROBE_DIR_MIDPROCESS}}/16S. To specify a directory outside \code{\link{NEONMICROBE_DIR_MIDPROCESS}}/16S, use the 'in_explicitdir' argument.}

\item{meta}{The output of \code{\link{downloadSequenceMetadata}}. Must be provided as either the data.frame returned by \code{\link{downloadSequenceMetadata}} or as a filepath to the csv file produced by \code{\link{downloadSequenceMetadata}}.}

\item{out_seqtab, out_track}{File locations where copies of the sequence table and read-tracking table (respectively) will be written. By default (NULL), these are \code{\link{NEONMICROBE_DIR_MIDPROCESS}}/16S/3_seqtabs/asv_16s_\link{timestamp}.Rds and \code{\link{NEONMICROBE_DIR_TRACKREADS}}/16S/dada_16s_\link{timestamp}.csv. If no copy should be saved, set the corresponding argument to FALSE.}

\item{in_explicitdir}{Directory name to use instead of 'in_subdir', if static directory name or directory outside of \code{\link{NEONMICROBE_DIR_MIDPROCESS}}/16S is desired. Not recommended for use within processing batches.}

\item{remove_chimeras}{(Optional) Default TRUE. Whether to remove chimeras from the sequence table. Currently only supports removal using \code{\link[dada2]{removeBimeraDenovo}} with the consensus method.}

\item{multithread}{Default FALSE. Whether to use multithreading.}

\item{verbose}{Default FALSE. Whether to print messages regarding the samples being processed, dimensions of the resulting sequence table, and the distribution of sequence lengths.}

\item{seed}{(Optional) Integer to use as random seed for reproducibility.}

\item{nbases}{(Optional) Number of bases to use for learning errors. Default 1e7.}
}
\value{
(Invisibly) A list of two elements. \strong{seqtab} is the sequence table, with chimeras removed if remove_chimeras == TRUE, and \strong{track} is a data frame displaying the number of reads remaining for each sample at various points throughout the processing pipeline.
}
\description{
Runs the Dada algorithm to infer sample composition from paired-end 16S fastq files.
This implementation is based on Ben Callahan's vignette at \url{https://benjjneb.github.io/dada2/bigdata_paired.html}.
}
\examples{
\dontrun{
seqtab.list <- runDada16S(c("sample1_R1.fastq", "sample1_R2.fastq", "sample2_R1.fastq", "sample2_R2.fastq"), meta, seed=1010100)
}
}
