% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_management.R
\name{setBatch}
\alias{setBatch}
\title{Switch to Existing Processing Batch}
\usage{
setBatch(
  batch_id = NULL,
  batches_dir = NULL,
  suppress_load_parameters = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{batch_id}{Character string. Unique ID of an existing processing batch.}

\item{batches_dir}{Directory where batch-specific directories are saved. By default (NULL), looks in NEONMICROBE_DIR_OUTPUTS(), ignoring batch modifications.}

\item{suppress_load_parameters}{Default FALSE. If TRUE, does not attempt to load parameters into the batch environment. Intended for use in newBatch()}

\item{verbose}{Default TRUE. Whether to print message confirming that batch has been set.}
}
\value{

}
\description{
Switch to Existing Processing Batch
}
