% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_wrappers.R
\name{trimPrimers16S}
\alias{trimPrimers16S}
\title{Trim Primers from 16S Sequences}
\usage{
trimPrimers16S(
  fn,
  dir_in,
  dir_out,
  primer_16S_fwd,
  primer_16S_rev,
  multithread = MULTITHREAD,
  post_samplename_pattern1 = "_R1.*\\\\.fastq",
  post_samplename_pattern2 = "_R2.*\\\\.fastq"
)
}
\arguments{
\item{fn}{Names of input fastq files, excluding directory path which is specified by dir_in. Files that do not exist will be ignored; however, if all files do not exist, this function will issue a warning.}

\item{dir_in}{Directory containing input fastq files.}

\item{dir_out}{Output directory. If it does not exist, it will be created.}

\item{primer_16S_fwd, primer_16S_rev}{DNA sequences of 16S forward and reverse primer, respectively}

\item{multithread}{Default MULTITHREAD in params.R. Whether to use multithreading. Note that Windows does not support multithreading in this function because it uses mclapply, so this argument must be set to FALSE on Windows systems.}

\item{post_samplename_pattern1, post_samplename_pattern2}{(Optional) Character pattern within the filename which immediately follows the end of the sample name. Defaults to "_R(1|2).*\\.fastq", as NEON fastq files typically consist of a sample name followed by "_R1.fastq" or "_R2.fastq", etc.}
}
\value{
Integer matrix denoting the number of reads remaining after primer-trimming for each input file.
}
\description{
Trims primers from 16S sequences using \code{\link[dada2]{filterAndTrim}}. This function
assumes that each read begins with its full primer sequence and operates
by truncating the beginning of each read by the length of its primer
sequence.
}
\examples{
\dontrun{
trimPrimers16S(c("sample1_R1.fastq", "sample1_R2.fastq", "sample2_R1.fastq", "sample2_R2.fastq"), "path/to/input", "path/to/output", "CCTACGGGNBGCASCAG", "GACTACNVGGGTATCTAATCC", multithread = TRUE)
}
}
