% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dada_wrappers.R
\name{trimPrimersITS}
\alias{trimPrimersITS}
\title{Trim Primers from ITS Sequences}
\usage{
trimPrimersITS(
  fn,
  dir_in,
  dir_out,
  primer_ITS_fwd,
  primer_ITS_rev,
  cutadapt_path = CUTADAPT_PATH,
  post_samplename_pattern1 = "_R1.*\\\\.fastq",
  post_samplename_pattern2 = "_R2.*\\\\.fastq",
  very_verbose = FALSE,
  discard_untrimmed = FALSE
)
}
\arguments{
\item{fn}{Names of input fastq files, excluding directory path which is specified by dir_in. Files that do not exist will be ignored; however, if all files do not exist, this function will issue a warning. It is assumed that these are R1 (forward-read) files only.}

\item{dir_in}{Directory containing input fastq files.}

\item{dir_out}{Output directory. If it does not exist, it will be created.}

\item{primer_ITS_fwd, primer_ITS_rev}{DNA sequence of the ITS forward and reverse primer, respectively.}

\item{cutadapt_path}{Default CUTADAPT_PATH in params.R. Path to cutadapt on your file system.}

\item{post_samplename_pattern1, post_samplename_pattern2}{(Optional) Character pattern within the filename which immediately follows the end of the sample name. Defaults to "_R(1|2).*\\.fastq", as NEON fastq files typically consist of a sample name followed by "_R1.fastq" or "_R2.fastq", etc.}

\item{very_verbose}{Default FALSE. Whether to print output from cutadapt. Unlike some other "verbose" arguments associated with the functions in this package, this does not default to VERBOSE in params.R.}

\item{discard_untrimmed}{Default FALSE. Whether to discard reads where a primer could not be found, leaving only those reads in which a primer has been trimmed.}
}
\value{
No value is returned.
}
\description{
Trims primers from ITS sequences using cutadapt. Cutadapt must be installed in order for this to work. Currently only supports R1 (forward-read) files.
}
\examples{
\dontrun{
trimPrimersITS(c("sample1_ITS_R1.fastq", "sample2_ITS_R1.fastq"), "path/to/input", "path/to/output", "CTTGGTCATTTAGAGGAAGTAA")
}
}
